some string utils and fixed math utils

+++ src/main/java/com/zipcodewilmington/danny_do_better_exercises/MathUtilities.java
@@ -20,7 +20,7 @@ public class MathUtilities {
      * @return sum of `baseValue` and `difference`
      */
     public Long add(long baseValue, long difference) {
-        return (long)(baseValue + difference);
+        return (baseValue + difference);
     }
 
     /**
@@ -47,7 +47,7 @@ public class MathUtilities {
      * @return sum of `baseValue` and `difference`
      */
     public Float add(float baseValue, float difference) {
-        return (float)(baseValue + difference);
+        return (baseValue + difference);
     }
 
     /**
@@ -56,7 +56,7 @@ public class MathUtilities {
      * @return sum of `baseValue` and `difference`
      */
     public Double add(double baseValue, double difference) {
-        return (double)(baseValue + difference);
+        return (baseValue + difference);
     }
 
     /**
@@ -74,7 +74,7 @@ public class MathUtilities {
      * @return difference between `baseValue` and `difference`
      */
     public Long subtract(long baseValue, long difference) {
-        return (long)(baseValue - difference);
+        return (baseValue - difference);
     }
 
     /**
@@ -101,7 +101,7 @@ public class MathUtilities {
      * @return difference between `baseValue` and `difference`
      */
     public Float subtract(float baseValue, float difference) {
-        return (float)(baseValue - difference);
+        return (baseValue - difference);
     }
 
     /**
@@ -110,7 +110,7 @@ public class MathUtilities {
      * @return difference between `baseValue` and `difference`
      */
     public Double subtract(double baseValue, double difference) {
-        return (double)(baseValue - difference);
+        return (baseValue - difference);
     }
 
 
@@ -129,7 +129,7 @@ public class MathUtilities {
      * @return division of `dividend` by `divisor
      */
     public Long divide(long dividend, long divisor) {
-        return (long)(dividend / divisor);
+        return (dividend / divisor);
     }
 
     /**
@@ -156,7 +156,7 @@ public class MathUtilities {
      * @return division of `dividend` by `divisor
      */
     public Float divide(float dividend, float divisor) {
-        return (float)(dividend / divisor);
+        return (dividend / divisor);
     }
 
     /**
@@ -165,7 +165,7 @@ public class MathUtilities {
      * @return division of `dividend` by `divisor
      */
     public Double divide(double dividend, double divisor) {
-        return (double)(dividend / divisor);
+        return (dividend / divisor);
     }
 
 
@@ -184,7 +184,7 @@ public class MathUtilities {
      * @return product of `multiplicand` by `multiplier`
      */
     public Long multiply(long multiplicand, long multiplier) {
-        return (long)(multiplicand * multiplier);
+        return (multiplicand * multiplier);
     }
 
     /**
@@ -210,7 +210,7 @@ public class MathUtilities {
      * @return product of `multiplicand` by `multiplier`
      */
     public Float multiply(float multiplicand, float multiplier) {
-        return (float)(multiplicand * multiplier);
+        return (multiplicand * multiplier);
     }
 
     /**
@@ -219,6 +219,6 @@ public class MathUtilities {
      * @return product of `multiplicand` by `multiplier`
      */
     public Double multiply(double multiplicand, double multiplier) {
-        return (double)(multiplicand * multiplier);
+        return (multiplicand * multiplier);
     }
 }
